feat: apply redis caching for fetching single post, like unlike, add comment and delete comment

+++ src/post/post.controller.ts
@@ -9,8 +9,10 @@ import {
   UseGuards,
   Req,
   NotFoundException,
+  Inject,
 } from '@nestjs/common';
 import { ApiBearerAuth, ApiTags } from '@nestjs/swagger';
+import { CACHE_MANAGER, CacheStore } from '@nestjs/cache-manager';
 
 import { PostService } from './post.service';
 import { CreatePostDto } from './dto/create-post.dto';
@@ -23,8 +25,10 @@ import { CommentPostDto } from './dto/comment-post.dto';
 @ApiBearerAuth()
 @ApiTags('Post')
 export class PostController {
-  constructor(private readonly postService: PostService) {}
-
+  constructor(
+    private readonly postService: PostService,
+    @Inject(CACHE_MANAGER) private cacheManager: CacheStore,
+  ) {}
   @Post()
   create(@Req() request: any, @Body() createPostDto: CreatePostDto) {
     const userId = request?.user?.id;
@@ -39,11 +43,19 @@ export class PostController {
 
   @Get(':id')
   async findOne(@Param('id') id: string) {
+    const cachedPost: any = await this.cacheManager.get(`post:${id}`);
+
+    if (cachedPost) {
+      return cachedPost;
+    }
+
     const post = await this.postService.findOne(id);
 
     if (!post) {
       throw new NotFoundException('Post not found');
     }
+
+    await this.cacheManager.set(`post:${id}`, post);
 
     return post;
   }
@@ -67,10 +79,15 @@ export class PostController {
   }
 
   @Delete(':id')
-  remove(@Req() request: any, @Param('id') id: string) {
+  async remove(@Req() request: any, @Param('id') id: string) {
     const userId = request?.user?.id;
 
-    return this.postService.remove(id, userId);
+    const [_deletedCached, deletedPost] = await Promise.all([
+      this.cacheManager.del(`post:${id}`),
+      this.postService.remove(id, userId),
+    ]);
+
+    return deletedPost;
   }
 
   @Post('like/:id')
@@ -82,8 +99,13 @@ export class PostController {
     }
 
     const userId = request?.user?.id;
+    const userLike = await this.postService.like(post.id, userId);
 
-    return this.postService.like(post.id, userId);
+    /* Updated Redis Cache */
+    post.Like.push(userLike);
+    await this.cacheManager.set(`post:${post.id}`, post);
+
+    return userLike;
   }
 
   @Delete('like/:id')
@@ -95,8 +117,13 @@ export class PostController {
     }
 
     const userId = request?.user?.id;
+    const userUnliked = await this.postService.unlike(post.id, userId);
 
-    return this.postService.unlike(post.id, userId);
+    /* Updated Redis Cache */
+    post.Like = post.Like.filter((userLike) => userLike.userId !== userId);
+    await this.cacheManager.set(`post:${post.id}`, post);
+
+    return userUnliked;
   }
 
   @Post('comment')
@@ -104,15 +131,46 @@ export class PostController {
     @Req() request: any,
     @Body() commentPostDto: CommentPostDto,
   ) {
-    const userId = request?.user?.id;
+    const post = await this.findOne(commentPostDto.postId);
 
-    return await this.postService.addComment(userId, commentPostDto);
+    if (!post) {
+      throw new NotFoundException('Post not found');
+    }
+
+    const userId = request?.user?.id;
+    const userComment = await this.postService.addComment(
+      userId,
+      commentPostDto,
+    );
+
+    const { userId: _userId, ...updatedComment } = userComment;
+
+    /* Updated Redis Cache */
+    post.Comment.push(updatedComment);
+    await this.cacheManager.set(`post:${post.id}`, post);
+
+    return userComment;
   }
 
   @Delete('comment/:id')
-  async deleteComment(@Req() request: any, @Param('id') id: string) {
-    const userId = request?.user?.id;
+  async deleteComment(
+    @Req() request: any,
+    @Param('id') id: string,
+    @Body() { postId }: { postId: string },
+  ) {
+    const post = await this.findOne(postId);
 
-    return await this.postService.deleteComment(userId, id);
+    if (!post) {
+      throw new NotFoundException('Post not found');
+    }
+
+    const userId = request?.user?.id;
+    const deletedComment = await this.postService.deleteComment(userId, id);
+
+    /* Updated Redis Cache */
+    post.Comment = post.Comment.filter((comment) => comment.id !== id);
+    await this.cacheManager.set(`post:${postId}`, post);
+
+    return deletedComment;
   }
 }
